soa-payments/requirements.txt
aiofiles==0.5.0
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
constantly==15.1.0
crochet==1.12.0
cryptography==2.9
cssselect==1.1.0
Cython==0.29.17
defusedxml==0.6.0
fastapi==0.54.0
h11==0.9.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
json5==0.9.4
jwt==1.0.0
lxml==4.5.0
MarkupSafe==1.1.1
netifaces==0.10.9
parsel==1.5.2
pip==20.0.2
Protego==0.1.16
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybreaker==0.6.0
pycparser==2.20
pycurl==7.43.0.5
pydantic==1.4
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyquery==1.4.1
python-consul==1.1.0
python-dateutil==2.8.1
python-dotenv==0.12.0
PyYAML==5.3.1
queuelib==1.5.0
requests==2.23.0
Scrapy==2.0.1
service-identity==18.1.0
setuptools==46.1.3
six==1.14.0
SQLAlchemy==1.3.15
starlette==0.13.2
stripe==2.45.0
tblib==1.6.0
tornado==4.5.3
Twisted==20.3.0
u-msgpack-python==2.5.2
urllib3==1.25.8
uvicorn==0.11.3
w3lib==1.21.0
websockets==8.1
Werkzeug==1.0.1
wheel==0.34.2
wincertstore==0.2
wrapt==1.12.1
WsgiDAV==3.0.3
zope.interface==5.1.0
